fix: divide landing page to smaller components
